component_video/pyproject.toml
[tool.poetry]
name = "video-frame-capture"
version = "0.1.0"
description = "Desktop application for video frame capture and export"
authors = ["Video Frame Capture Team <team@example.com>"]
packages = [{include = "video_frame_capture", from = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
PySide6 = "^6.6.0"
opencv-python = "^4.8.0"
numpy = "^1.24.0"
Pillow = "^10.0.0"
cv2PySide6 = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-qt = "^4.2.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[tool.poetry.group.build.dependencies]
pyinstaller = "^6.4.0"
briefcase = "^0.3.16"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
video-frame-capture = "video_frame_capture.main:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true